Implement InputColumnsManager to keep track of input & selectable column properties
