+++ mathesar_ui/src/systems/query-builder/output-config-sidebar/OutputConfigSidebar.svelte
@@ -1,0 +1,20 @@
+<script lang="ts">
+  import type QueryManager from '../QueryManager';
+
+  export let queryManager: QueryManager;
+</script>
+
+<div class="output-config-sidebar">
+</div>
+
+<style lang="scss">
+  .output-config-sidebar {
+    width: 22rem;
+    border-left: 1px solid #efefef;
+    flex-shrink: 0;
+    flex-grow: 0;
+    flex-basis: 22rem;
+    display: flex;
+    flex-direction: column;
+  }
+</style>
